remove unused index.html, move example fetch to own folder
