--- index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-    <meta name="keywords" content="p5.js, p5.js web editor, web editor, processing, code editor" />
-    <meta name="description" content="A web editor for sketching using p5.js library." />
-    <title>p5.js Web Editor</title>
-    <link href='https://fonts.googleapis.com/css?family=Inconsolata' rel='stylesheet' type='text/css'>
-    <link href='https://fonts.googleapis.com/css?family=Montserrat:400,700' rel='stylesheet' type='text/css'>
-  </head>
-  <body>
-    <div id="root" class="root-app">
-    </div>
-    <script src="/dist/vendor.js"></script>
-    <script src="/dist/app.js"></script>
-  </body>
-</html>
