+++ server/scripts/fetch-examples-gg.js
@@ -1,4 +1,4 @@
 require('babel-register');
 require('babel-polyfill');
 require('dotenv').config();
-require('./server/examples.js');
+require('./examples-gg-latest.js');
