+++ server/scripts/fetch-examples.js
@@ -1,4 +1,4 @@
 require('babel-register');
 require('babel-polyfill');
 require('dotenv').config();
-require('./server/examples-gg-latest.js');
+require('./examples.js');
